[Upd] Bring back getting output columns back to the TabularModule from TabularReader because it doesn't require reading the actual file.
