+++ s-pipes-modules/module-tabular/src/main/java/cz/cvut/spipes/modules/TabularModule.java
@@ -10,6 +10,7 @@ import cz.cvut.spipes.constants.SML;
 import cz.cvut.spipes.engine.ExecutionContext;
 import cz.cvut.spipes.engine.ExecutionContextFactory;
 import cz.cvut.spipes.exception.ResourceNotFoundException;
+import cz.cvut.spipes.exception.ResourceNotUniqueException;
 import cz.cvut.spipes.modules.annotations.SPipesModule;
 import cz.cvut.spipes.modules.exception.SheetDoesntExistsException;
 import cz.cvut.spipes.modules.exception.SheetIsNotSpecifiedException;
@@ -271,7 +272,18 @@ public class TabularModule extends AnnotatedAbstractModule {
             em.close();
             em.getEntityManagerFactory().close();
 
-            outputColumns = tabularReader.getOutputColumns(header);
+            Set<String> columnNames = new HashSet<>();
+
+            for (String columnTitle : header) {
+                String columnName = normalize(columnTitle);
+                boolean isDuplicate = !columnNames.add(columnName);
+
+                Column schemaColumn = new Column(columnName, columnTitle);
+
+                outputColumns.add(schemaColumn);
+                schemaColumn.setTitle(columnTitle);
+                if (isDuplicate) throwNotUniqueException(schemaColumn, columnTitle, columnName);
+            }
 
             for(Column schemaColumn : outputColumns){
                 String columnName = schemaColumn.getName();
@@ -375,6 +387,18 @@ public class TabularModule extends AnnotatedAbstractModule {
         return false;
     }
 
+    private void throwNotUniqueException(Column column, String columnTitle, String columnName) {
+        throw new ResourceNotUniqueException(
+                String.format("Unable to create value of property %s due to collision. " +
+                                "Both column titles '%s' and '%s' are normalized to '%s' " +
+                                "and thus would refer to the same property url <%s>.",
+                        CSVW.propertyUrl,
+                        columnTitle,
+                        column.getTitle(),
+                        columnName,
+                        column.getPropertyUrl()));
+    }
+
     private TableSchema getTableSchema(EntityManager em) {
         TypedQuery<TableSchema> query = em.createNativeQuery(
             "PREFIX csvw: <http://www.w3.org/ns/csvw#>\n" +
@@ -404,6 +428,10 @@ public class TabularModule extends AnnotatedAbstractModule {
         } else {
             return ExecutionContextFactory.createContext(JenaUtils.createUnion(inputModel, outputModel));
         }
+    }
+
+    private String normalize(String label) {
+        return label.trim().replaceAll("[^\\w]", "_");
     }
 
     @Override
