+++ s-pipes-modules/module-tabular/src/main/java/cz/cvut/spipes/modules/tabular/TabularReader.java
@@ -12,6 +12,5 @@ import java.util.Set;
 
 public interface TabularReader {
     List<String> getHeader() throws IOException;
-    List<Column> getOutputColumns(List<String>header);
     List<Statement> getRowStatements(List<String>header, List<Column>outputColumns, TableSchema tableSchema) throws IOException;
 }
